Restructured files to meet best practices
