--- Lernjournal/src/app/Eintrag.ts
@@ -1,5 +1,0 @@
-export interface Eintrag{
-  titel: string;
-  text: string;
-  id: number;
-}
