+++ Lernjournal/src/app/app.module.ts
@@ -1,46 +1,27 @@
-import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
-import { AppComponent } from './app.component';
-import { EditComponent } from './edit/edit.component';
-import { MainComponent } from './main/main.component';
+import {NgModule} from '@angular/core';
+import {BrowserModule} from '@angular/platform-browser';
+import {AppComponent} from './app.component';
+import {EditComponent} from "./features/edit/edit.component";
+import {MainComponent} from "./features/main/main.component";
 import {AppRoutingModule} from "./app-routing-module";
 import {HttpClientModule} from "@angular/common/http";
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
-import { EintragComponent } from './eintrag/eintrag.component';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import {EintragComponent} from "./features/eintrag/eintrag.component";
+import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {MatIconModule} from "@angular/material/icon";
 import {MatListModule} from "@angular/material/list";
 import {MatInputModule} from "@angular/material/input";
 import {MatButtonModule} from "@angular/material/button";
 import {MatCheckboxModule} from "@angular/material/checkbox";
 import {MatTooltipModule} from "@angular/material/tooltip";
-import { RegistrationComponent } from './registration/registration.component';
-import { LoginComponent } from './login/login.component';
+import {RegistrationComponent} from "./features/registration/registration.component";
+import {LoginComponent} from "./features/login/login.component";
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    EditComponent,
-    MainComponent,
-    EintragComponent,
-    RegistrationComponent,
-    LoginComponent,
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    HttpClientModule,
-    FormsModule,
-    BrowserAnimationsModule,
-    MatIconModule,
-    MatListModule,
-    MatInputModule,
-    MatButtonModule,
-    MatCheckboxModule,
-    ReactiveFormsModule,
-    MatTooltipModule
-  ],
+  declarations: [AppComponent, EditComponent, MainComponent, EintragComponent, RegistrationComponent, LoginComponent,],
+  imports: [BrowserModule, AppRoutingModule, HttpClientModule, FormsModule, BrowserAnimationsModule, MatIconModule, MatListModule, MatInputModule, MatButtonModule, MatCheckboxModule, ReactiveFormsModule, MatTooltipModule],
   providers: [],
   bootstrap: [AppComponent]
 })
-export class AppModule { }
+export class AppModule {
+}
