+++ Lernjournal/src/app/core/pdf-export.service.ts
