+++ Lernjournal/src/app/features/edit/edit.component.css
