+++ Lernjournal/src/app/features/eintrag/eintrag.component.html
