+++ Lernjournal/src/app/features/eintrag/eintrag.component.spec.ts
