+++ Lernjournal/src/app/features/main/main.component.html
