+++ Lernjournal/src/app/features/registration/registration.component.css
