+++ Lernjournal/src/app/features/registration/registration.component.html
