+++ Lernjournal/src/app/features/registration/registration.component.spec.ts
