+++ Lernjournal/src/app/features/registration/registration.component.ts
@@ -1,6 +1,6 @@
 import {Component} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
-import {RegistrationService} from "../registration.service";
+import {RegistrationService} from "../../core/registration.service";
 
 @Component({
   selector: 'app-registration',
