--- Lernjournal/src/app/pdf-export.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { PdfExportService } from './pdf-export.service';
-
-describe('PdfExportService', () => {
-  let service: PdfExportService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(PdfExportService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
