+++ Lernjournal/src/app/types/eintrag.ts
@@ -1,0 +1,5 @@
+export interface Eintrag {
+  id: number;
+  titel: string;
+  text: string;
+}
