+++ Lernjournal/src/app/types/registration.ts
@@ -1,4 +1,4 @@
-export interface Registration{
+export interface Registration {
   id: number;
   email: string;
   password: string;
